remove kind and namespace parameters from namelabelmatch to support CREATE requests
The 'input.review' object doesn't contain the 'kind.kind' or 'namespace' fields in CREATE requests
because these details belong to existing resources, whereas 'input.review.object' does because
this represents the object to be created.

When we tried to assign 'input.review.kind.kind' and 'input.review.namespace' to variables in CREATE requests,
it would evaluate to false before even testing the condition, meaning this constraint would never produce a
violation for a CREATE request.

We were only grabbing these values to export them as 'details' and this information is already output by kube when
a request is blocked by gatekeeper, so the best thing to do is to just remove them.

With this change, the constraint now works for newly created resources.