use yq, rather than a go program, to check the template src rego
The golang image used by Dockerfile.constraint is pretty large
and makes the test take an unreasonable amount of time to complete
when you don't already have the image pulled.

With yq, the image is alpine based so it's about 30x smaller and it
also means we aren't maintaining our own yaml parsing code.